+++ packages/blurry-image-lazy-loading/e2e/common.e2e-spec.ts
@@ -1,5 +1,4 @@
 import { AssetServerPlugin } from "@vendure/asset-server-plugin";
-import { AssetImporter } from "@vendure/core";
 import { createTestEnvironment } from "@vendure/testing";
 import path from "path";
 import { afterAll, assert, beforeAll, describe, test } from "vitest";
@@ -8,8 +7,10 @@ import { testConfig } from "../../../utils/e2e/test-config";
 import { ThumbHashStrategy } from "../src/config/ThumbHashStrategy";
 import { PluginPreviewImageHashResultCode as CODE } from "../src/generated-admin-types";
 import { PreviewImageHashPlugin } from "../src/preview-image-hash.plugin";
-import { CREATE_FOR_COLLECTION, CREATE_FOR_PRODUCT, CREATE_PREVIEW_IMG_HASH } from "./graphql/admin-e2e-definitions";
+import { CREATE_FOR_ALL_ASSETS, CREATE_FOR_COLLECTION, CREATE_FOR_PRODUCT, CREATE_PREVIEW_IMG_HASH } from "./graphql/admin-e2e-definitions";
 import {
+  CreateForAllAssetsMutation,
+  CreateForAllAssetsMutationVariables,
   CreateForCollectionMutation,
   CreateForCollectionMutationVariables,
   CreateForProductMutation,
@@ -18,7 +19,7 @@ import {
   CreatePreviewImageHashMutationVariables,
 } from "./types/generated-admin-types";
 
-describe("General API", () => {
+describe("General API", { concurrent: true }, () => {
   const { server, adminClient, shopClient } = createTestEnvironment({
     ...testConfig(8002),
     importExportOptions: {
@@ -35,17 +36,17 @@ describe("General API", () => {
     ],
   });
 
+  // This collection holds 3 "plant"-products via facets, see last items in import csv
+  const idCollection = "2";
+
   beforeAll(async () => {
     await server.init({
       productsCsvPath: path.join(__dirname, "../../../utils/e2e/e2e-products-full.csv"),
       initialData: initialData,
       customerCount: 2,
+      logging: true,
     });
     await adminClient.asSuperAdmin();
-
-    const fixturesAssets = [/* T_1 */ "vendure-brand-icon-2024-primary.jpeg"];
-
-    await server.app.get(AssetImporter).getAssets(fixturesAssets);
   }, 60000);
 
   afterAll(async () => {
@@ -53,46 +54,39 @@ describe("General API", () => {
   });
 
   test("Successfully enqueue hashing for collection with default arguments", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreateForCollectionMutation, CreateForCollectionMutationVariables>(
       CREATE_FOR_COLLECTION,
-      { input: { idCollection: "1" } },
+      { input: { idCollection, regenerateExistingHashes: true } },
     );
 
     expect(result.pluginPreviewImageHashCreateImageHashesForCollection.code).toBe(CODE.OK);
-    // TODO would be good to seed a more practical env
-    expect(result.pluginPreviewImageHashCreateImageHashesForCollection.jobsAddedToQueue).toStrictEqual(0);
+    expect(result.pluginPreviewImageHashCreateImageHashesForCollection.jobsAddedToQueue).toStrictEqual(3);
     expect(result.pluginPreviewImageHashCreateImageHashesForCollection.assetsSkipped).toStrictEqual(0);
   });
 
   test("Successfully enqueue hashing for collection with batch size zero", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreateForCollectionMutation, CreateForCollectionMutationVariables>(
       CREATE_FOR_COLLECTION,
-      { input: { idCollection: "1", batchSize: 0 } },
+      { input: { idCollection, regenerateExistingHashes: true, batchSize: 0 } },
     );
 
     expect(result.pluginPreviewImageHashCreateImageHashesForCollection.code).toBe(CODE.OK);
-    // TODO would be good to seed a more practical env
-    expect(result.pluginPreviewImageHashCreateImageHashesForCollection.jobsAddedToQueue).toStrictEqual(0);
+    expect(result.pluginPreviewImageHashCreateImageHashesForCollection.jobsAddedToQueue).toStrictEqual(3);
     expect(result.pluginPreviewImageHashCreateImageHashesForCollection.assetsSkipped).toStrictEqual(0);
   });
 
   test("Successfully enqueue hashing for collection with negative batch size", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreateForCollectionMutation, CreateForCollectionMutationVariables>(
       CREATE_FOR_COLLECTION,
-      { input: { idCollection: "1", batchSize: -1 } },
+      { input: { idCollection, regenerateExistingHashes: true, batchSize: -1 } },
     );
 
     expect(result.pluginPreviewImageHashCreateImageHashesForCollection.code).toBe(CODE.OK);
-    // TODO would be good to seed a more practical env
-    expect(result.pluginPreviewImageHashCreateImageHashesForCollection.jobsAddedToQueue).toStrictEqual(0);
+    expect(result.pluginPreviewImageHashCreateImageHashesForCollection.jobsAddedToQueue).toStrictEqual(3);
     expect(result.pluginPreviewImageHashCreateImageHashesForCollection.assetsSkipped).toStrictEqual(0);
   });
 
   test("Successfully add task to job queue", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreatePreviewImageHashMutation, CreatePreviewImageHashMutationVariables>(
       CREATE_PREVIEW_IMG_HASH,
       {
@@ -110,7 +104,6 @@ describe("General API", () => {
   });
 
   test("Successfully enqueue job for product", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreateForProductMutation, CreateForProductMutationVariables>(
       CREATE_FOR_PRODUCT,
       {
@@ -120,13 +113,57 @@ describe("General API", () => {
 
     assert(result.pluginPreviewImageHashCreateImageHashesForProduct.__typename === "PluginPreviewImageHashResult");
     expect(result.pluginPreviewImageHashCreateImageHashesForProduct.code).toBe(CODE.OK);
-    // TODO would be good to seed a more practical env
-    expect(result.pluginPreviewImageHashCreateImageHashesForProduct.jobsAddedToQueue).toStrictEqual(0);
+    expect(result.pluginPreviewImageHashCreateImageHashesForProduct.jobsAddedToQueue).toStrictEqual(1);
     expect(result.pluginPreviewImageHashCreateImageHashesForProduct.assetsSkipped).toStrictEqual(0);
   });
 
-  test("Fail to enqueue job for non-existent product", async ({ expect }) => {
+  test("Successfully enqueue jobs for hashing all assets", async ({ expect }) => {
+    const result = await adminClient.query<CreateForAllAssetsMutation, CreateForAllAssetsMutationVariables>(
+      CREATE_FOR_ALL_ASSETS,
+      { input: { regenerateExistingHashes: true } }
+    );
+
+    assert(result.pluginPreviewImageHashCreateImageHashesForAllAssets.__typename === "PluginPreviewImageHashResult");
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.code).toBe(CODE.OK);
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.jobsAddedToQueue).toStrictEqual(4);
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.assetsSkipped).toStrictEqual(1); // Skipped SVG
+  });
+
+  test("Successfully enqueue hashing for all assets with negative batch size", async ({ expect }) => {
+    const result = await adminClient.query<CreateForAllAssetsMutation, CreateForAllAssetsMutationVariables>(
+      CREATE_FOR_ALL_ASSETS,
+      { input: { batchSize: -1, regenerateExistingHashes: true } },
+    );
+
+    assert(result.pluginPreviewImageHashCreateImageHashesForAllAssets.__typename === "PluginPreviewImageHashResult");
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.code).toBe(CODE.OK);
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.jobsAddedToQueue).toStrictEqual(4);
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.assetsSkipped).toStrictEqual(1); // Skipped SVG
+  });
+
+  test("Successfully skip hashing for all assets due to existing hashes", async ({ expect }) => {
+    await adminClient.query<CreateForAllAssetsMutation, CreateForAllAssetsMutationVariables>(
+      CREATE_FOR_ALL_ASSETS,
+      { input: { regenerateExistingHashes: true } },
+    );
+
+    // Wait a little so that the server can hash the images
+    await new Promise((res, rej) => setTimeout(res, 1000));
     await adminClient.asSuperAdmin();
+
+    const result = await adminClient.query<CreateForAllAssetsMutation, CreateForAllAssetsMutationVariables>(
+      CREATE_FOR_ALL_ASSETS,
+      { input: { regenerateExistingHashes: false } },
+    );
+
+    assert(result.pluginPreviewImageHashCreateImageHashesForAllAssets.__typename === "PluginPreviewImageHashResult");
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.code).toBe(CODE.OK);
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.jobsAddedToQueue).toStrictEqual(0);
+    expect(result.pluginPreviewImageHashCreateImageHashesForAllAssets.assetsSkipped).toStrictEqual(5);
+  });
+
+
+  test("Fail to enqueue job for non-existent product", async ({ expect }) => {
     const result = await adminClient.query<CreateForProductMutation, CreateForProductMutationVariables>(
       CREATE_FOR_PRODUCT,
       {
@@ -141,7 +178,6 @@ describe("General API", () => {
   });
 
   test("Fail to enqueue hashing for collection due to non-existing collection id", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreateForCollectionMutation, CreateForCollectionMutationVariables>(
       CREATE_FOR_COLLECTION,
       { input: { idCollection: "non-existent-collection-id" } },
@@ -153,7 +189,6 @@ describe("General API", () => {
   });
 
   test("Fail to generate hash due to non-existent asset via api", async ({ expect }) => {
-    await adminClient.asSuperAdmin();
     const result = await adminClient.query<CreatePreviewImageHashMutation, CreatePreviewImageHashMutationVariables>(
       CREATE_PREVIEW_IMG_HASH,
       {
@@ -189,7 +224,7 @@ describe("General API", () => {
     await adminClient.asAnonymousUser();
     await expect(
       adminClient.query<CreateForCollectionMutation, CreateForCollectionMutationVariables>(CREATE_FOR_COLLECTION, {
-        input: { idCollection: "1" },
+        input: { idCollection },
       }),
     ).rejects.toThrowError("You are not currently authorized to perform this action");
   });