+++ packages/blurry-image-lazy-loading/src/api/api-extensions.ts
@@ -44,6 +44,25 @@ export const adminApiExtensions = gql`
     height: Int
   }
 
+  input PluginPreviewImageHashForAllAssetsInput {
+    """
+    How large a page is when paginating through all of the assets. Increasing this number will minimize 
+    needed roundtrips between Vendure and the database. Be careful though, this increases load on the database.
+
+    @default 50
+    """
+    batchSize: Int
+
+    """
+    Assets can be quite numerous, so by default this mutation skips over assets that already have a hash set, in order to minimize the needed compute.
+    This can be undesirable, for example after you change the strategy, the encoding or resize options.
+    Setting this option to \`true\` will overwrite the existing hashes, letting you update your existing assets.
+
+    @default false
+    """
+    regenerateExistingHashes: Boolean
+  }
+
   input PluginPreviewImageHashForCollectionInput {
     """
     ID of a collection
@@ -95,6 +114,14 @@ export const adminApiExtensions = gql`
     ): PluginPreviewImageHashCreateResult!
 
     """
+    Create preview image hashes for a product.
+    This includes both the product itself and all of its ProductVariant assets.
+    """
+    pluginPreviewImageHashCreateImageHashesForProduct(
+      input: PluginPreviewImageHashForProductInput!
+    ): PluginPreviewImageHashResult!
+
+    """
     Create preview image hashes for an entire collection.
     This includes the collection, the contained Product-assets and related ProductVariant-assets.
 
@@ -110,11 +137,13 @@ export const adminApiExtensions = gql`
     ): PluginPreviewImageHashResult!
 
     """
-    Create preview image hashes for a product.
-    This includes both the product itself and all of its ProductVariant assets.
+    Create preview image hashes for all assets.
+    
+    This mutation should be handled with extra care since an installation may hold hundreds of thousands of images.
+    This is mainly useful as a one-time-use utility to initialize all of the assets with hashes after installing the plugin.
     """
-    pluginPreviewImageHashCreateImageHashesForProduct(
-      input: PluginPreviewImageHashForProductInput!
+    pluginPreviewImageHashCreateImageHashesForAllAssets(
+      input: PluginPreviewImageHashForAllAssetsInput
     ): PluginPreviewImageHashResult!
-  }
+}
 `;