+++ packages/blurry-image-lazy-loading/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 (2025-06-11)
+
+- Add new mutation for hashing all assets called `pluginPreviewImageHashCreateImageHashesForAllAssets`
+- Update README
+- Internal: add & improve test suite
+
 ## 1.0.1 (2025-04-02)
 
 - Add example in README for generating DataUrl in browsers